Refactored css styles into app.modules.css
