+++ public/locales/en/translation.json
@@ -852,7 +852,7 @@
     "title": "Settings",
     "general": "General",
     "actionItemCategories": "Action Item Categories",
-    "updateOrganization": "Update Organization",
+    "updateOrganization": "Edit Organization",
     "seeRequest": "See Request",
     "noData": "No data",
     "otherSettings": "Other Settings",
