+++ src/components/OrgSettings/General/DeleteOrg/DeleteOrg.module.css
@@ -1,4 +1,4 @@
-.settingsBody {
+/* .settingsBody {
   margin: 2.5rem 0;
 }
 
@@ -22,4 +22,4 @@
 .cardBody .textBox {
   margin: 0 0 3rem 0;
   color: var(--bs-secondary);
-}
+} */
