+++ src/components/OrgSettings/General/OrgProfileFieldSettings/OrgProfileFieldSettings.module.css
@@ -1,4 +1,4 @@
-.customDataTable {
+/* .customDataTable {
   width: 100%;
   border-collapse: collapse;
 }
@@ -21,4 +21,4 @@ form {
 .saveButton {
   width: 10em;
   align-self: self-end;
-}
+} */
