+++ src/components/OrgSettings/General/OrgUpdate/OrgUpdate.module.css
@@ -1,4 +1,4 @@
-.message {
+/* .message {
   height: 420px;
   display: flex;
   justify-content: center;
@@ -10,4 +10,4 @@
   transform: scale(1.5);
   color: var(--bs-danger);
   margin-bottom: 1rem;
-}
+} */
